S3C-35 introduce a customized multilevel module and client/server wrappers
DISCLAIMER: temporary solution, a rewrite of our own multilevel is
planned. Out for review as-is to spot early issues.

The new fork of multilevel is to be reworked to a new independent
implementation later. The changes allow to use sublevel properly with
it, to allow the client to create new sublevels without the
disconnect/manifest update/etc. burden. It does not allow anymore
custom functions per sublevel.

The public API using the multilevel fork as a backend is located in
arsenal.storage.metadata.(server|client)

NOTE: S3 uses this new API in branch wip/S3C-35-bucketFileRework.